[INDY-1138] Increment plenum version, rename module
